% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056_mod.R
\name{getDirectCouplingCost}
\alias{getDirectCouplingCost}
\title{Direct Coupling Costs}
\usage{
getDirectCouplingCost(techno.struct)
}
\arguments{
\item{techno.struct}{Design structure matrix.}
}
\value{
numeric value of of direct coupling costs
}
\description{
Calculates the direct coupling costs of a design structure matrix.
}
