% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056_mod.R
\name{getExperiment}
\alias{getExperiment}
\title{Experimental Design}
\usage{
getExperiment(
  authority.v,
  ticks,
  composability.v,
  m.v,
  f.v,
  turbulence.v,
  release.v,
  size = 100,
  repetitions,
  seed,
  plot = FALSE
)
}
\arguments{
\item{authority.v}{Vector of logical values. Determines the mode of coordination. If TRUE then authority-based coordination is used.}

\item{ticks}{Numeric. The number of time steps running the simulation.}

\item{composability.v}{Vector of numeric values. Sets the degree of contribution interdependence between 0 and 1.}

\item{m.v}{Vector of numeric values. Determines the number of contribution created in each time step.}

\item{release.v}{Vector of numeric values. Sets the delay of visibility measured in time steps.}

\item{size}{Width (and breadth) of the design structure matrix to be created in the simulation. Default: 100.}

\item{repetitions}{Numeric. Number of repetition for each parameter setting.}

\item{seed}{Numeric. Value for the random seed of the first experiment.}

\item{plot}{Logical. If TRUE then a plot is created displaying key measures of the design structure matrix during the simulation run ; default: FALSE}
}
\value{
Dataframe of simulation runs.
}
\description{
The function getExperiment() allows to run experiments over parameter spaces. The function calls the run() function for each parameter combination and writes the simulation results in a dataframe.
}
\examples{
getExperiment(authority.v=c(TRUE), ticks=1000, composability.v=c(0.070), m.v=c(100),f.v=c(1.00), turbulence.v=c(0), release.v=c(1), size=100, repetitions=1, seed=28071984, plot=FALSE)
getExperiment(authority.v=c(FALSE), ticks=1000, composability.v=c(0.070), m.v=c(100),f.v=c(1.00), turbulence.v=c(0), release.v=c(1), size=100, repetitions=1, seed=28071984, plot=FALSE)
getExperiment(authority.v=c(TRUE), ticks=1000, composability.v=c(seq(0.000,0.050,0.005)), m.v=c(100), f.v=c(1.00), turbulence.v=c(0), release.v=c(1), size=100, repetitions=1, seed=28071984, plot=FALSE)
}
