% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056_mod.R
\name{getPropagationCost}
\alias{getPropagationCost}
\title{Propagation Costs}
\usage{
getPropagationCost(techno.struct)
}
\arguments{
\item{techno.struct}{Design structure matrix.}
}
\value{
numeric value of propagation costs
}
\description{
Calculates the propagation costs of a design structure matrix.
}
\details{
Sources: MacCormack, A., Rusnak, J., Baldwin, C.Y., 2006. Exploring the Structure of Complex Software Designs: An Empirical Study of Open Source and Proprietary Code. Management Science 52(7).

Sharman, D., Yassine, A., 2004. Characterizing complex product architectures. Systems Engineering Journal 7(1).
}
