% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056_mod.R
\name{getTopdown.v}
\alias{getTopdown.v}
\title{Identifying Required Contributions}
\usage{
getTopdown.v(
  sol.focal.pos,
  tec.struct,
  sol.struct,
  sol.functs,
  req.v = vector(mode = "numeric", length = length(sol.functs))
)
}
\arguments{
\item{sol.focal.pos}{Numeric. The position of the focal contribution for which requirements are searched for.}

\item{tec.struct}{The current design structure matrix.}

\item{sol.struct}{Dataframe containing the requirement vectors the contributions.}

\item{sol.functs}{Vector containing the functions of all contributions.}

\item{req.v}{Vector marking the required contributions.}
}
\description{
Identifying Required Contributions
}
