% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056_mod.R
\name{getVisibilityMatrix}
\alias{getVisibilityMatrix}
\title{Visibility Matrix}
\usage{
getVisibilityMatrix(techno.struct, N)
}
\arguments{
\item{techno.struct}{Design structure matrix.}

\item{N}{Width (or breadth) of the design structure matrix.}
}
\value{
Visibility matrix as a matrix object.
}
\description{
Creates a visibility matrix from a design structure matrix.
}
\details{
Source: Sharman, D., Yassine, A., 2004. Characterizing complex product architectures. Systems Engineering Journal 7(1).
}
