% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056_mod.R
\name{run}
\alias{run}
\title{Simulation Run}
\usage{
run(
  authority,
  ticks,
  composability,
  m = 1,
  f = 1,
  turbulence = 0,
  release = 1,
  size = 100,
  plot = T
)
}
\arguments{
\item{authority}{logical. Determines the mode of coordination. If TRUE then authority-based coordination is used.}

\item{ticks}{Numeric. The number of time steps running the simulation.}

\item{composability}{Numeric. Sets the degree of contribution interdependence between 0 and 1.}

\item{m}{Numeric. Determines the number of contribution created in each time step.}

\item{release}{Numeric. Sets the delay of visibility measured in time steps. Default: 1 (i.e. no release delay).}

\item{size}{Width (and breadth) of the design structure matrix to be created in the simulation. Default: 100.}

\item{plot}{Logical. If TRUE then a plot is created displaying key measures of the design structure matrix during the simulation run ; default: FALSE}
}
\value{
List containing measures of the simulation run.
}
\description{
Single simulation run of the simulation model.
}
